Refactor functions to accept values.
Refactored functions to accept the values they need rather than the full
data object. This encourages storing intermediate results in variables
to avoid recalculating values such as BMR multiple times to obtain the
goal TDEE. This means a little more work for users who only want part of
the data (initial TDEE) because they must be explicit about calculating
the BMR, but it reduces the amount of processing done.

This will be alleviated for users who want the full breakdown with the
calculate() function which will hide this complexity.